branching parser WIP 2: Electric Boogaloo
Now with checkpoints and branches

New token structure (it makes everything just work better) and new lexer to go with it
The Lookups class now contains crimes against programming. Don't look at it. (I don't know the difference between a function and an object)
Changed parser. Its not any better, just different.

parserFunction and parserFunctionNew are in the process of merging.
added checkpoints and a branch manager. They work, but only barely.