Branches, like a tree
Like a tree, we have branches and birds.

Changes:
Berd made a new friend, Deno.
Updated the ReadMe with all the features we'll want to add.
Added tests to test code that needs testing using tests.

New token structure and new lexer to go with it. Tokens now contain all possible interpretation of the source from their starting position. When creating tokens, the lexer will try to create a token in every possible way, and only keep valid interpretations.

Changed the parser. It now creates branches for every possible way to parse the source code. This way, we don't have to know what interpretation we will need, we can just parse all of them and chose at runtime. It will make the runtime a bit more complex and a lot slower, but it will allow us to parse any kind of syntax, even if they are ambiguous or requires context from the rest of the program.

The ParsingFunctionMaps class now contains crimes against programming. Don't look at it.